left company and update to appointment create -- require email

+++ custom/modules/Calls/logic_hooks/leftcompany.php
@@ -5,12 +5,24 @@ class leftCompany{
 			$contact = BeanFactory::getBean('Contacts', $bean->parent_id);
 			if(!empty($contact)){
 				if(!empty($contact->account_id)){
-					$account = BeanFactory::getBean('Accounts', $contact->account_id);
-					//$GLOBALS['log']->fatal($contact->id);
-					//$GLOBALS['log']->fatal($account->id);
-					$account->load_relationship('contacts');
-					$account->contacts->delete($acount->id,$contact->id);
-					//$account->save();
+				
+                                	$account = BeanFactory::getBean('Accounts', $contact->account_id);
+				
+					//$account->load_relationship('contacts');
+					//$account->contacts->delete($acount->id,$contact->id);
+
+                                        if($contact->left_company_c==1){
+                                            // checkbox is checked
+                                           $account->load_relationship('contacts');
+                                           $account->contacts->delete($acount->id,$contact->id);
+                                           $contact->left_company_c=0;
+                                        }else{
+                                           // checkbox is not checked
+                                           $contact->left_company_c=1;
+                                        
+                                        }
+                                        $contact->save();
+					
 				}
 			}
 		}
